adding cart , wishlist , and some of configration for internationalization for a language

+++ src/main/java/com/stackdeans/loginregister/models/Category.java
@@ -10,7 +10,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", unique = true, nullable = false)
     private String name;
@@ -28,17 +28,17 @@ public class Category {
     }
 
 
-    public Category(Integer id, String name, String picture, List<Category> children) {
+    public Category(Long id, String name, String picture, List<Category> children) {
         this.id = id;
         this.name = name;
         this.picture = picture;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
